Move resolve_aliases test to test_backend module. Remove test_data_types module.
