+++ test/test_backend.py
@@ -5,6 +5,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import unittest
 
 from stone.ir import (
+    Alias,
     ApiNamespace,
     ApiRoute,
     List,
@@ -12,6 +13,7 @@ from stone.ir import (
     String,
     Struct,
     StructField,
+    resolve_aliases
 )
 from stone.backend import CodeBackend
 
@@ -225,6 +227,25 @@ int sq(int x) <
         t = _TesterCmdline(None, ['-v'])
         self.assertTrue(t.args.verbose)
 
+    def test_resolve_aliases(self):
+        first_alias = Alias(None, None, None)
+        first_alias.data_type = String()
+        returned_alias = resolve_aliases(first_alias)
+
+        # Test that single-level alias chain resolves
+        self.assertEqual(returned_alias, first_alias)
+        self.assertIsInstance(returned_alias.data_type, String)
+
+        first_alias = Alias(None, None, None)
+        second_alias = Alias(None, None, None)
+        first_alias.data_type = second_alias
+        second_alias.data_type = String()
+
+        # Test that a two-level alias chain resolves
+        returned_alias = resolve_aliases(first_alias)
+        self.assertEqual(returned_alias, first_alias)
+        self.assertIsInstance(first_alias.data_type, String)
+        self.assertIsInstance(second_alias.data_type, String)
 
 if __name__ == '__main__':
     unittest.main()
